@@ -19,24 +19,24 @@
 		"@fortawesome/svelte-fontawesome": "^0.2.2",
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/adapter-node": "^5.2.9",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
+		"@sveltejs/kit": "^2.5.27",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
 		"@types/eslint": "^8.56.7",
 		"autoprefixer": "^10.4.19",
 		"eslint": "^9.0.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.36.0",
+		"eslint-plugin-svelte": "^2.45.1",
 		"flowbite": "^2.5.2",
 		"flowbite-svelte": "^0.47.2",
 		"flowbite-svelte-icons": "^1.6.2",
 		"globals": "^15.0.0",
 		"postcss": "^8.4.38",
 		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
+		"prettier-plugin-svelte": "^3.2.6",
 		"prettier-plugin-tailwindcss": "^0.6.4",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0",
 		"tailwindcss": "^3.4.4",
-		"vite": "^5.0.3",
+		"vite": "^5.4.4",
 		"vitest": "^1.2.0"
 	},
 	"type": "module",
