@@ -1,10 +1,10 @@
 <script>
 	import { createEventDispatcher } from 'svelte';
-	export let route;
+	let { route } = $props();
 	const dispatch = createEventDispatcher();
 
 	function handleClick() {
-		dispatch('routeClick', { route });
+		// dispatch('routeClick', { route })
 	}
 
 	function getDisplayRouteName() {
@@ -21,7 +21,7 @@
 <button
 	type="button"
 	class="route-item flex w-full items-center justify-between border-b border-gray-200 bg-[#f9f9f9] p-4 text-left hover:bg-[#e9e9e9] focus:outline-none dark:border-[#313135] dark:bg-[#1c1c1c] dark:text-white dark:hover:bg-[#363636]"
-	on:click={handleClick}
+	onclick={handleClick}
 >
 	<div class="text-lg font-semibold" style="color: #{route.color}">
 		{getDisplayRouteName(route)}
