@@ -1,6 +1,6 @@
 <script>
 	import { createEventDispatcher } from 'svelte';
-	export let stop;
+	let { stop } = $props();
 	const dispatch = createEventDispatcher();
 
 	function handleClick() {
@@ -11,7 +11,7 @@
 <button
 	type="button"
 	class="stop-item dark:bg[#000000] flex w-full items-center justify-between border-b border-gray-200 bg-[#f9f9f9] p-4 text-left hover:bg-[#e9e9e9] focus:outline-none dark:border-[#313135] dark:bg-[#1c1c1c] dark:hover:bg-[#363636]"
-	on:click={handleClick}
+	onclick={handleClick}
 >
 	<div class="text-lg font-semibold text-[#000000] dark:text-white">
 		{stop.name}
