@@ -2,7 +2,15 @@
 	import { getContext } from 'svelte';
 	import { slide } from 'svelte/transition';
 
-	export let data = null;
+	/**
+	 * @typedef {Object} Props
+	 * @property {any} [data]
+	 * @property {import('svelte').Snippet} [header]
+	 * @property {import('svelte').Snippet} [children]
+	 */
+
+	/** @type {Props} */
+	let { data = null, header, children } = $props();
 
 	const id = crypto.randomUUID();
 	const { registerItem } = getContext('accordion');
@@ -19,10 +27,10 @@
 			class="flex w-full items-center justify-between py-3 text-left text-base font-medium text-gray-500 hover:text-gray-900 dark:text-gray-400 dark:hover:text-white"
 			class:text-gray-900={$isActive}
 			class:dark:text-white={$isActive}
-			on:click={toggle}
+			onclick={toggle}
 			aria-expanded={$isActive}
 		>
-			<slot name="header" />
+			{@render header?.()}
 			<svg
 				class="h-6 w-6 shrink-0 transition-transform"
 				class:rotate-180={$isActive}
@@ -43,7 +51,7 @@
 			}}
 		>
 			<div class="py-3">
-				<slot />
+				{@render children?.()}
 			</div>
 		</div>
 	{/if}
