@@ -1,4 +1,6 @@
 <script>
+	import { run } from 'svelte/legacy';
+
 	import { browser } from '$app/environment';
 	import { createEventDispatcher, onMount, onDestroy } from 'svelte';
 	import {
@@ -13,19 +15,30 @@
 	import { faBus } from '@fortawesome/free-solid-svg-icons';
 	import { RouteType, routePriorities, prioritizedRouteTypeForDisplay } from '$config/routeConfig';
 	import { isMapLoaded } from '$src/stores/mapStore';
+	/**
+	 * @typedef {Object} Props
+	 * @property {any} [selectedTrip]
+	 * @property {any} [selectedRoute]
+	 * @property {boolean} [showRoute]
+	 * @property {boolean} [showRouteMap]
+	 * @property {any} [mapProvider]
+	 */
 
-	export let selectedTrip = null;
-	export let selectedRoute = null;
-	export let showRoute = false;
-	export let showRouteMap = false;
-	export let mapProvider = null;
+	/** @type {Props} */
+	let {
+		selectedTrip = null,
+		selectedRoute = null,
+		showRoute = false,
+		showRouteMap = false,
+		mapProvider = null
+	} = $props();
 
-	let isTripPlanModeActive = false;
+	let isTripPlanModeActive = $state(false);
+	let mapInstance = $state(null);
+	let mapElement = $state();
+	let allStops = $state([]);
 
-	let mapInstance = null;
-	let mapElement;
 	let markers = [];
-	let allStops = [];
 	let stopsCache = new Map();
 
 	const dispatch = createEventDispatcher();
@@ -142,19 +155,6 @@
 		}
 	}
 
-	$: {
-		if (selectedRoute) {
-			clearAllMarkers();
-			updateMarkers();
-		} else if (!isTripPlanModeActive) {
-			allStops.forEach((s) => addMarker(s));
-		}
-	}
-
-	$: if (isTripPlanModeActive) {
-		clearAllMarkers();
-	}
-
 	function addMarker(s) {
 		if (!mapInstance) {
 			console.error('Map not initialized yet');
@@ -230,6 +230,19 @@
 				element.parentNode.removeChild(element);
 			}
 		});
+	});
+	run(() => {
+		if (selectedRoute) {
+			clearAllMarkers();
+			updateMarkers();
+		} else if (!isTripPlanModeActive) {
+			allStops.forEach((s) => addMarker(s));
+		}
+	});
+	run(() => {
+		if (isTripPlanModeActive) {
+			clearAllMarkers();
+		}
 	});
 </script>
 
