@@ -15,7 +15,7 @@
 	import ModalPane from '$components/navigation/ModalPane.svelte';
 	import StopPane from '$components/stops/StopPane.svelte';
 
-	export let stop;
+	let { stop } = $props();
 </script>
 
 <ModalPane on:close title={stop.name}>
