@@ -2,9 +2,7 @@
 	import LegDetails from './LegDetails.svelte';
 	import { formatTime } from '$lib/formatters';
 	import { t } from 'svelte-i18n';
-	export let itinerary;
-	export let expandedSteps;
-	export let toggleSteps;
+	let { itinerary, expandedSteps, toggleSteps } = $props();
 </script>
 
 <div class="mb-4 flex items-center justify-between rounded-lg bg-gray-100 p-4 shadow-md">
