@@ -1,6 +1,12 @@
 <script>
-	export let text = 'From';
-	export let color = '#79aa38';
+	/**
+	 * @typedef {Object} Props
+	 * @property {string} [text]
+	 * @property {string} [color]
+	 */
+
+	/** @type {Props} */
+	let { text = 'From', color = '#79aa38' } = $props();
 </script>
 
 <div
